Implement fsck (#46)
This fsck travels into all directories in a filesystem dcrtime database
and ensures that all hashes and timestamps are consistent. It also
verifies that flushed records are in fact stored on the blockchain and
verifies the merkle roots in the flush record and reality.

It currently fixes data errors caused by three bugs. These bugs have
since been fixed but the data incoherence was never corrected.

The fsck has a journaling function which records all actions taken. This
is useful in case at a later time there needs to be a correction.