api/v2: Duplicate merkle branch definition
This removes the dependency on the main module's merkle package.

The JSON marshaling of the merkle branch is disgusting but this would need to
be cleaned up in a v3 API if anyone attempts this.

Part of a larger change to make actually make the api/v2 submodule actually
go-gettable.  It is completely busted at the moment, due to importing the main
module's merkle package without a require on the main module.